pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "workflow-backend"
version = "0.1.0"
description = "Sistema de orquestación de workflows con IA (API-first, FastAPI)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "sqlmodel>=0.0.21",
  "SQLAlchemy>=2.0.30",
  "alembic>=1.13.1",
  "redis>=5.0.4",
  "httpx>=0.27.0",
  "python-ulid>=2.7.0",
  "orjson>=3.10.0",
  "prometheus-client>=0.20.0",
  "python-multipart>=0.0.9",
  "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.8",
  "schemathesis>=3.27.6",
  "ruff>=0.6.9",
  "black>=24.8.0",
  "mypy>=1.10.0",
  "bandit>=1.7.9",
  "pip-audit>=2.7.3",
  "types-redis",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E","F","I","N","UP","B","SIM"]
ignore = []
exclude = ["alembic"]

[tool.pytest.ini_options]
pythonpath = ["app"]
asyncio_mode = "auto"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["alembic*", "docker*", "spec*", "tests*"]

